Consolidate gnxi tests into gnmi directory (#22553)
* Consolidate gnxi tests into gnmi directory

Move gnoi tests from tests/gnxi to tests/gnmi and remove redundant
coverage. The gnxi tests use the cleaner ptf_gnoi fixture approach
which is now preferred over the old gnoi_request helper and native
gRPC stubs.

Changes:
- Move test_gnoi_file.py from gnxi to gnmi
- Replace test_gnoi_system.py with gnxi version using ptf_gnoi fixture
- Remove test_gnoi_system_grpc.py (redundant System.Time test)
- Remove grpc_utils.py (no longer needed)
- Delete tests/gnxi directory


* Fix stale gnxi references after consolidation into gnmi

Update testcases.json and pr_reviewer-by-files.yml to reflect
the gnxi -> gnmi test directory migration merged from master.


* just to rerun pipeline


* Revert auto-populated reviewer file to fix pre-commit


---------

(cherry picked from commit 3657cce706b8efcba25c7d119ad54a8a57401b62)
