+++ tests/gnmi/conftest.py
@@ -1,18 +1,14 @@
 import pytest
 import logging
-import os
-import glob
-import grpc
-
-from grpc_tools import protoc
 
 from tests.common.helpers.assertions import pytest_require as pyrequire
 from tests.common.helpers.dut_utils import check_container_state
 from tests.gnmi.helper import gnmi_container, apply_cert_config, recover_cert_config
 from tests.gnmi.helper import GNMI_SERVER_START_WAIT_TIME, check_ntp_sync_status
 from tests.common.gu_utils import create_checkpoint, rollback
-from tests.common.helpers.gnmi_utils import GNMIEnvironment, create_revoked_cert_and_crl, \
-                                            create_gnmi_certs, delete_gnmi_certs, create_ext_conf
+from tests.common.helpers.gnmi_utils import create_revoked_cert_and_crl, create_gnmi_certs, \
+    delete_gnmi_certs, prepare_root_cert, prepare_server_cert, prepare_client_cert, copy_certificate_to_dut, \
+    copy_certificate_to_ptf
 from tests.common.helpers.ntp_helper import setup_ntp_context
 
 
@@ -82,81 +78,12 @@ def setup_gnmi_rotated_server(duthosts, rand_one_dut_hostname, localhost, ptfhos
         check_container_state(duthost, gnmi_container(duthost), should_be_running=True),
         "Test was not supported on devices which do not support GNMI!"
     )
-
-    # Create Root key
-    local_command = "openssl genrsa -out gnmiCA.key 2048"
-    localhost.shell(local_command)
-
-    # Create Root cert
-    local_command = "openssl req \
-                        -x509 \
-                        -new \
-                        -nodes \
-                        -key gnmiCA.key \
-                        -sha256 \
-                        -days 1825 \
-                        -subj '/CN=test.gnmi.sonic' \
-                        -out gnmiCA.pem"
-    localhost.shell(local_command)
-
-    # Create server key
-    local_command = "openssl genrsa -out gnmiserver.key 2048"
-    localhost.shell(local_command)
-
-    # Create server CSR
-    local_command = "openssl req \
-                        -new \
-                        -key gnmiserver.key \
-                        -subj '/CN=test.server.gnmi.sonic' \
-                        -out gnmiserver.csr"
-    localhost.shell(local_command)
-
-    # Sign server certificate
-    create_ext_conf(duthost.mgmt_ip, "extfile.cnf")
-    local_command = "openssl x509 \
-                        -req \
-                        -in gnmiserver.csr \
-                        -CA gnmiCA.pem \
-                        -CAkey gnmiCA.key \
-                        -CAcreateserial \
-                        -out gnmiserver.crt \
-                        -days 825 \
-                        -sha256 \
-                        -extensions req_ext -extfile extfile.cnf"
-    localhost.shell(local_command)
-
-    # Create client key
-    local_command = "openssl genrsa -out gnmiclient.key 2048"
-    localhost.shell(local_command)
-
-    # Create client CSR
-    local_command = "openssl req \
-                        -new \
-                        -key gnmiclient.key \
-                        -subj '/CN=test.client.gnmi.sonic' \
-                        -out gnmiclient.csr"
-    localhost.shell(local_command)
-
-    # Sign client certificate
-    local_command = "openssl x509 \
-                        -req \
-                        -in gnmiclient.csr \
-                        -CA gnmiCA.pem \
-                        -CAkey gnmiCA.key \
-                        -CAcreateserial \
-                        -out gnmiclient.crt \
-                        -days 825 \
-                        -sha256"
-    localhost.shell(local_command)
-
-    create_revoked_cert_and_crl(localhost, ptfhost, duthost)
-
-    # Copy CA certificate, server certificate and client certificate over to the DUT
-    duthost.copy(src='gnmiCA.pem', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiserver.crt', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiserver.key', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiclient.crt', dest='/etc/sonic/telemetry/')
-    duthost.copy(src='gnmiclient.key', dest='/etc/sonic/telemetry/')
+    prepare_root_cert(localhost)
+    prepare_server_cert(duthost, localhost)
+    prepare_client_cert(localhost)
+    copy_certificate_to_ptf(ptfhost)
+    create_revoked_cert_and_crl(localhost, ptfhost)
+    copy_certificate_to_dut(duthost)
 
 
 @pytest.fixture(scope="module")
@@ -175,91 +102,3 @@ def check_dut_timestamp(duthosts, rand_one_dut_hostname, localhost):
     time_diff = local_time - dut_time
     if time_diff >= GNMI_SERVER_START_WAIT_TIME:
         logger.warning("DUT time is wrong (%d), please check NTP" % (-time_diff))
-
-
-def compile_protos(proto_files, proto_root):
-    """Compile all .proto files using grpc_tools.protoc."""
-    for proto_file in proto_files:
-
-        # Command arguments for protoc
-        args = [
-            "grpc_tools.protoc",
-            f"--proto_path={proto_root}",  # Root directory for proto imports
-            f"--python_out={proto_root}",     # Output for message classes
-            f"--grpc_python_out={proto_root}",  # Output for gRPC stubs
-            proto_file                     # Input .proto file
-        ]
-
-        print(f"Compiling: {proto_file}")
-        ret_code = protoc.main(args)
-        if ret_code != 0:
-            raise Exception(f"Failed to compile {proto_file} with return code {ret_code}")
-
-
-def cleanup_generated_files():
-    """Remove all generated proto .py files."""
-    generated_files = glob.glob("gnmi/protos/**/*.py")
-    for file in generated_files:
-        os.remove(file)
-
-
-@pytest.fixture(scope="module")
-def setup_and_cleanup_protos():
-    """Compile proto files before running tests and remove them afterward."""
-    PROTO_ROOT = "gnmi/protos"
-    PROTO_FILES = ["gnmi/protos/gnoi/system/system.proto"]
-
-    # Compile proto files into Python gRPC stubs
-    compile_protos(PROTO_FILES, PROTO_ROOT)
-
-    # Run tests, then clean up
-    yield
-    cleanup_generated_files()
-
-
-@pytest.fixture(scope="function")
-def grpc_channel(duthosts, rand_one_dut_hostname):
-    """
-    Fixture to set up a gRPC channel with secure credentials.
-    """
-    duthost = duthosts[rand_one_dut_hostname]
-
-    # Get DUT gRPC server address and port
-    if ":" in duthost.mgmt_ip and not duthost.mgmt_ip.startswith('['):
-        ip = f"[{duthost.mgmt_ip}]"
-    else:
-        ip = duthost.mgmt_ip
-    env = GNMIEnvironment(duthost, GNMIEnvironment.GNMI_MODE)
-    port = env.gnmi_port
-    target = f"{ip}:{port}"
-
-    # Load the TLS certificates
-    with open("gnmiCA.pem", "rb") as f:
-        root_certificates = f.read()
-    with open("gnmiclient.crt", "rb") as f:
-        client_certificate = f.read()
-    with open("gnmiclient.key", "rb") as f:
-        client_key = f.read()
-
-    # Create SSL credentials
-    credentials = grpc.ssl_channel_credentials(
-        root_certificates=root_certificates,
-        private_key=client_key,
-        certificate_chain=client_certificate,
-    )
-
-    # Create gRPC channel
-    logging.info("Creating gRPC secure channel to %s", target)
-    channel = grpc.secure_channel(target, credentials)
-
-    try:
-        grpc.channel_ready_future(channel).result(timeout=10)
-        logging.info("gRPC channel is ready")
-    except grpc.FutureTimeoutError as e:
-        logging.error("Error: gRPC channel not ready: %s", e)
-        pytest.fail("Failed to connect to gRPC server")
-
-    yield channel
-
-    # Close the channel
-    channel.close()
