+++ tests/gnmi/test_gnoi_system.py
@@ -1,56 +1,29 @@
+"""
+Simple integration tests for gNOI System service.
+
+All tests automatically run with TLS server configuration by default.
+Users don't need to worry about TLS configuration.
+"""
 import pytest
 import logging
-import json
 
-from .helper import gnoi_request
-from tests.common.helpers.assertions import pytest_assert
-import re
+# Import fixtures to ensure pytest discovers them
+from tests.common.fixtures.grpc_fixtures import (  # noqa: F401
+    setup_gnoi_tls_server, ptf_gnoi, ptf_grpc
+)
 
+logger = logging.getLogger(__name__)
+
+# Enable TLS fixture by default for all tests in this module
 pytestmark = [
     pytest.mark.topology('any'),
-    pytest.mark.usefixtures("setup_gnmi_ntp_client_server", "setup_gnmi_server",
-                            "setup_gnmi_rotated_server", "check_dut_timestamp")
+    pytest.mark.usefixtures("setup_gnoi_tls_server")
 ]
 
 
-"""
-This module contains tests for the gNOI System API.
-"""
-
-
-def test_gnoi_system_time(duthosts, rand_one_dut_hostname, localhost):
-    """
-    Verify the gNOI System Time API returns the current system time in valid JSON format.
-    """
-    duthost = duthosts[rand_one_dut_hostname]
-
-    # Get current time
-    ret, msg = gnoi_request(duthost, localhost, "System", "Time", "")
-    pytest_assert(ret == 0, "System.Time API reported failure (rc = {}) with message: {}".format(ret, msg))
-    logging.info("System.Time API returned msg: {}".format(msg))
-    # Message should contain a json substring like this {"time":1735921221909617549}
-    # Extract JSON part from the message
-    msg_json = extract_first_json_substring(msg)
-    if not msg_json:
-        pytest.fail("Failed to extract JSON from System.Time API response")
-    logging.info("Extracted JSON: {}".format(msg_json))
-    pytest_assert("time" in msg_json, "System.Time API did not return time")
-
-
-def extract_first_json_substring(s):
-    """
-    Extract the first JSON substring from a given string.
-
-    :param s: The input string containing JSON substring.
-    :return: The first JSON substring if found, otherwise None.
-    """
-
-    json_pattern = re.compile(r'\{.*?\}')
-    match = json_pattern.search(s)
-    if match:
-        try:
-            return json.loads(match.group())
-        except json.JSONDecodeError:
-            logging.error("Failed to parse JSON: {}".format(match.group()))
-            return None
-    return None
+def test_system_time(ptf_gnoi):  # noqa: F811
+    """Test System.Time RPC with TLS enabled by default."""
+    result = ptf_gnoi.system_time()
+    assert "time" in result
+    assert isinstance(result["time"], int)
+    logger.info(f"System time: {result['time']} nanoseconds since epoch")
